提交请假条、审批请假条

+++ app/src/main/java/com/zhaoweihao/architechturesample/ui/LoginActivity.java
@@ -195,7 +195,7 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
             com.zhaoweihao.architechturesample.database.User user3 = DataSupport.findLast(com.zhaoweihao.architechturesample.database.User.class);
             user3.delete();
             com.zhaoweihao.architechturesample.database.User user = new com.zhaoweihao.architechturesample.database.User();
-            user.setUserId(1000);
+            user.setStuId(1000);
             user.setUsername("zhaoweihao22");
             user.setStudentId("2015191054");
             user.setClassId("20151912");
@@ -235,11 +235,13 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
                 log(thisClass, "保存到数据body"+body);
                 //解析json数据组装RestResponse对象
                 RestResponse restResponse = new Gson().fromJson(body, RestResponse.class);
+
                 // 修复登录密码不正确异常
                 if (restResponse == null) {
                     runOnUiThread(() -> Toast.makeText(LoginActivity.this, "请检查密码", Toast.LENGTH_SHORT).show());
                     return;
                 }
+
                 if ( restResponse.getCode() == 500 ) {
                     log(thisClass, "登录失败，请检查用户名和密码");
                     try {
@@ -272,11 +274,14 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
 
 
                     try {
-                        com.zhaoweihao.architechturesample.database.User user3 = DataSupport.findLast(com.zhaoweihao.architechturesample.database.User.class);
-                        user3.delete();
-                        com.zhaoweihao.architechturesample.database.User user = new Gson().fromJson(restResponse.getPayload().toString(),com.zhaoweihao.architechturesample.database.User.class);
+                        List<com.zhaoweihao.architechturesample.database.User> allDatas = DataSupport.findAll(com.zhaoweihao.architechturesample.database.User.class);
+                        if(allDatas.size()>0){
+                            com.zhaoweihao.architechturesample.database.User user3 = DataSupport.findLast(com.zhaoweihao.architechturesample.database.User.class);
+                            user3.delete();
+                        }
+                        com.zhaoweihao.architechturesample.data.User user = new Gson().fromJson(restResponse.getPayload().toString(),com.zhaoweihao.architechturesample.data.User.class);
                         com.zhaoweihao.architechturesample.database.User user1 = new com.zhaoweihao.architechturesample.database.User();
-                        user1.setUserId(user.getUserId());
+                        user1.setStuId(user.getId());
                         user1.setUsername(user.getUsername());
                         user1.setStudentId(user.getStudentId());
                         user1.setClassId(user.getClassId());
@@ -290,7 +295,8 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
                         //保存到数据库
                         user1.save();
                         List<com.zhaoweihao.architechturesample.database.User> allNews = DataSupport.findAll(com.zhaoweihao.architechturesample.database.User.class);
-                        log(thisClass, "保存到数据库成功"+user.getUserId()+"当前所有的user个数"+allNews.size());
+
+                        log(thisClass, "保存到数据库成功"+user.getId()+"**********"+user1.getStuId()+"当前所有的user个数"+allNews.size());
                     } catch (Exception e) {
                         e.printStackTrace();
                     }
